requirements.txt
flask
numpy<2.0
pillow
scikit-learn
matplotlib
pandas
open_clip-torch